💄 design: 헤더 클릭 시 마이페이지 및 로그아웃 디자인만 추가
